requirements.txt
alembic==1.13.1
antlr4-python3-runtime==4.9.3
appnope==0.1.4
asttokens==2.4.1
blinker==1.7.0
blosc2==2.6.2
casadi==3.6.5
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==2.2.1
comm==0.2.2
contourpy==1.2.0
cues==0.3.0
cycler==0.12.1
Cython==3.0.10
databricks-cli==0.18.0
debugpy==1.8.1
decorator==5.1.1
dill==0.3.8
docker==6.1.3
entrypoints==0.4
executing==2.0.1
filelock==3.13.3
Flask==2.3.3
fonttools==4.50.0
fsspec==2024.3.1
gitdb==4.0.11
GitPython==3.1.42
gunicorn==20.1.0
h5py==3.11.0
idna==3.6
importlib-metadata==6.11.0
iniconfig==2.0.0
ipykernel==6.29.4
ipython==8.24.0
ipywidgets==8.1.2
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.3
joblib==1.3.2
jupyter_client==8.6.1
jupyter_core==5.7.2
jupyterlab_widgets==3.0.10
kiwisolver==1.4.5
Mako==1.3.2
Markdown==3.6
markdown-it-py==2.2.0
MarkupSafe==2.1.5
matplotlib==3.8.3
matplotlib-inline==0.1.7
matplotx==0.3.10
mdurl==0.1.2
mlflow==2.4.1
mpmath==1.3.0
msgpack==1.0.8
ndindex==1.8
nest-asyncio==1.6.0
networkx==3.2.1
numexpr==2.10.0
numpy==1.26.4
oauthlib==3.2.2
omegaconf==2.3.0
packaging==23.2
pandas==2.2.1
parso==0.8.4
pexpect==4.9.0
pillow==10.2.0
platformdirs==4.2.1
plotly==5.20.0
pluggy==1.5.0
prompt-toolkit==3.0.43
protobuf==4.25.3
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
py-cpuinfo==9.0.0
pyarrow==12.0.1
Pygments==2.17.2
PyJWT==2.8.0
pyparsing==3.1.2
PyQt5==5.15.10
PyQt5-Qt5==5.15.13
PyQt5-sip==12.13.0
pytest==7.4.4
python-dateutil==2.9.0.post0
pytz==2023.4
PyYAML==6.0.1
pyzmq==26.0.2
querystring-parser==1.2.4
regelum-control==0.3.2
rehydra==1.3.5
rehydra-joblib-launcher==1.3.0
requests==2.31.0
rich==13.3.1
scikit-learn==1.4.1.post1
scipy==1.12.0
six==1.16.0
smmap==5.0.1
SQLAlchemy==2.0.29
sqlparse==0.4.4
stack-data==0.6.3
svgpath2mpl==1.0.0
svgpathtools==1.6.1
svgwrite==1.4.3
sympy==1.12
tables @ git+https://github.com/PyTables/PyTables@0af6a199e6347268d00008900ce328feb0f91b8e
tabulate==0.9.0
tenacity==8.2.3
threadpoolctl==3.4.0
torch==2.2.1
tornado==6.4
tqdm==4.66.4
traitlets==5.14.3
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
wcwidth==0.2.13
websocket-client==1.7.0
Werkzeug==3.0.1
widgetsnbextension==4.0.10
zipp==3.18.1
